+++ glean-core/ios/Glean/Glean.swift
@@ -59,7 +59,7 @@ public class Glean {
     /// once.
     ///
     /// A LifecycleObserver will be added to submit pings when the application goes
-    /// into the background.
+    /// into the foreground and background.
     ///
     /// - parameters:
     ///     * uploadEnabled: A `Bool` that enables or disables telemetry.
@@ -333,6 +333,11 @@ public class Glean {
         return self.initialized
     }
 
+    /// Handle foreground event and submit appropriate pings
+    func handleForegroundEvent() {
+        Pings.shared.baseline.submit(reason: .foreground)
+    }
+
     /// Handle background event and submit appropriate pings
     func handleBackgroundEvent() {
         Pings.shared.baseline.submit(reason: .background)
@@ -379,7 +384,7 @@ public class Glean {
         }
 
         if !self.getUploadEnabled() {
-            self.logger.error("Glean must be enabled before sending pings")
+            self.logger.error("Glean disabled: not submitting any pings")
             return
         }