+++ glean-core/python/tests/test_network.py
@@ -8,6 +8,7 @@ import uuid
 
 from glean import Glean
 from glean.net import PingUploadWorker
+from glean.net.http_client import HttpClientUploader
 
 
 def test_invalid_filename():
@@ -102,10 +103,15 @@ def test_ping_upload_worker_single_process(safe_httpserver):
     p1 = PingUploadWorker._process()
     p2 = PingUploadWorker._process()
 
-    p1.join()
-    assert p1.exitcode == 0
+    p1.wait()
+    assert p1.returncode == 0
 
-    p2.join()
-    assert p2.exitcode == 0
+    p2.wait()
+    assert p2.returncode == 0
 
     assert 100 == len(safe_httpserver.requests)
+
+
+def test_unknown_url_no_exception():
+    # Shouldn't leak any socket or HTTPExceptions
+    assert not HttpClientUploader.upload("http://nowhere.example.com", "{}", [])