+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v27.1.0...master)
+[Full changelog](https://github.com/mozilla/glean/compare/v28.0.0...master)
+
+# v28.0.0 (2020-04-23)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v27.1.0...v28.0.0)
+
+* General:
+  * The baseline ping is now sent when the application goes to foreground, in addition to background and dirty-startup.
+* Python:
+  * BUGFIX: The ping uploader will no longer display a trace back when the upload fails due to a failed DNS lookup, network outage, or related issues that prevent communication with the telemetry endpoint.
+  * The dependency on `inflection` has been removed.
+  * The Python bindings now use `subprocess` rather than `multiprocessing` to perform ping uploading in a separate process. This should be more compatible on all of the platforms Glean supports.
 
 # v27.1.0 (2020-04-09)
 
@@ -17,7 +28,7 @@
 
 [Full changelog](https://github.com/mozilla/glean/compare/v26.0.0...v27.0.0)
 
-* General:
+* General
   * Glean will now detect when the upload enabled flag changes outside of the application, for example due to a change in a config file. This means that if upload is disabled while the application wasn't running (e.g. between the runs of a Python command using the Glean SDK), the database is correctly cleared and a deletion request ping is sent. See [#791](https://github.com/mozilla/glean/pull/791).
   * The `events` ping now includes a reason code: `startup`, `background` or `max_capacity`.
 * iOS: